software/lab/pdf-parser-manual/requirements.txt
pypdf2
pdfminer.six
bs4
python-docx==0.8.10
colorama 
termcolor
six
pyfiglet